+++ dlls/xaudio2_1/Makefile.in
@@ -13,6 +13,7 @@ C_SRCS = \
 	xaudio_voice_master.c \
 	xaudio_voice_wrap.c \
 	xaudio_callbacks.c \
+	xapo.c \
 	xapo_wrap.c \
 	FAudio/FAPOBase.c \
 	FAudio/FAudio.c \
